+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 ## 🆕 Recent Updates
 
-[![Backend Tests](https://github.com/mariusroyale/blackjack-game-ts/actions/workflows/backend-tests.yml/badge.svg)](https://github.com/mariusroyale/blackjack-game-ts/actions/workflows/backend-tests.yml)
+[![Backend Tests](https://github.com/mariusroyale/blackjack-game-ts/actions/workflows/backend-tests.yml/badge.svg)](https://github.com/mariusroyale/blackjack-game-ts/actions/workflows/backend-tests.yml) | [![Frontend Tests](https://github.com/mariusroyale/blackjack-game-ts/actions/workflows/frontend-tests.yml/badge.svg)](https://github.com/mariusroyale/blackjack-game-ts/actions/workflows/frontend-tests.yml)
 
 ### May 6, 2025
 - 🧪 Added **unit tests** for Gaming Service using jest testing suite
+- 🧪 Added **unit tests** for Frontend react app using vitest testing suite
 
 ### May 5, 2025
 - ♠️ Switched from 1-deck to a 6-deck shoe (300+ cards) for realistic gameplay